src/main/java/com/moo/cdn/model/Backend.java
package com.moo.cdn.model;

/**
 * https://docs.fastly.com/api/config#backend
 */
public class Backend {

    private String address;
    private String name;
    private String hostname;
    private Integer port;
    private Integer version;

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Integer getPort() {
        return port;
    }

    public void setPort(Integer port) {
        this.port = port;
    }

    public Integer getVersion() {
        return version;
    }

    public void setVersion(Integer version) {
        this.version = version;
    }

    public String getHostname() {
        return hostname;
    }

    public void setHostname(String hostname) {
        this.hostname = hostname;
    }
}
